NETOBSERV-649 fix incorrect metrics on "common" filters
The issue is described here: https://issues.redhat.com/browse/NETOBSERV-649?focusedCommentId=21091465&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-21091465

The solution implemented is, for common filters (eg. source=X OR
dest=X), to include a "not" match in the dest query: source=X OR
(source!=X AND dest=X) to eliminate the overlap.

It was needed to add this NOT operator on backend side (so half the way
done for NETOBSERV-387 which is about implementing NOT)

Also, the matrix-merger algorithm had to be updated so that values are
added when the same metric is returned by several queries. We can do
that precisely because we can now assume they don't overlap.