NETOBSERV-656 overview page changes and additions  (#210)
* NETOBSERV-649 fix incorrect metrics on "common" filters

The issue is described here: https://issues.redhat.com/browse/NETOBSERV-649?focusedCommentId=21091465&page=com.atlassian.jira.plugin.system.issuetabpanels%3Acomment-tabpanel#comment-21091465

The solution implemented is, for common filters (eg. source=X OR
dest=X), to include a "not" match in the dest query: source=X OR
(source!=X AND dest=X) to eliminate the overlap.

It was needed to add this NOT operator on backend side (so half the way
done for NETOBSERV-387 which is about implementing NOT)

Also, the matrix-merger algorithm had to be updated so that values are
added when the same metric is returned by several queries. We can do
that precisely because we can now assume they don't overlap.

* Fix broken k8sResource Filters

Refactored frontend filters encoding to help refining when the "not"
filter has to apply (rule of thumb is: it applies where there is a
possible overlap between src and dest metrics results, hence, when the
filter applies on objects that can have internal traffic (where src == dest)

* NETOBSERV-656 overview page changes and additions

- Remove the MetricFunction dropdown as it became almost useless after
  latest metrics updates
- Adapt Donuts so it uses fixed metricFunction (avg or sum, proportions
  are equivalent), and add a new donut for latest rate
- Use Total metric in donuts to show total;
- Allow to include the internal traffic in charts (internal means:
  src==dest)
- Add others in donuts
- Add controls to show/hide internal traffic and "others" in donuts
- Create new chart that shows topK and total on the same chart

Fix issues with tooltip/legend

Use kebab menu for chart options

Move kebab to header, better metric identification

Unknown metrics can now be roughly identified, depending on the scope,
as non-pod network or non-host network.
Add a new filter option in donuts/total-bars to filter out of scope
metrics

Move kebab to header (had to refactor for components interaction)