NETOBSERV-474 summary filters by source or dest
!QE NOTICE: graph node id generation changed!

Allow to filter by source, dest or common via the summary panel filters

This is also adding Owner as part of the information displayed per node.
It modifies the whole topology node data so that the entire
"TopologyMetricPeer" (returned from metrics) object is now part of it.

It also makes metrics matching for groups simpler because the parent information is now contained in the id itself

Also, some refactoring on element-panel to extract new components: element-field(s), peer-resource-link, summary-filter-button