NETOBSERV-1269 refactor overlapping detection for BNF
This fix allows working on situations as described in
https://issues.redhat.com/browse/NETOBSERV-1269 and, hopefully, is a
more generic / consistent solution anyway.

Changing how overlapping results are managed: work on results rather
than on query to remove overlapping parts.

When fetching flows, we don't explicitly remove the overlaps, but this
is managed implicitly with the existing dedup functions

When fetching metrics, we consider that the computed metrics for any
label set is the ORIGINAL metrics + the SWAPPED metrics - the OVERLAP
metrics.

Add tests, fix tests

Note that adjustment is necessary in MetricsQuerySummary test as the
"endWithTolerance" calculation has changed a little bit, modifying
computed rates ... new values are still close to the previous ones

Rename callback function