NETOBSERV-739: Add prometheus queries
This introduces prometheus as a new datasource, that can be used either
as a replacement or as a complement of Loki. It doesn't require to
change the frontend/backend API interface: on every frontend query,
backend checks if that query is transposable to prometheus, and if so,
runs it on prometheus, else falls back on loki (if it's enabled)

There's some refactoring of the config and topology handlers to make
place for prometheus.