NETOBSERV-2441: fix page loading triggering autocompletion (#1060)
There was an abusive use of the autocompletion mechanism in filters.
It's primarily used to get a list of filter options out of a search value.
But it was also used, during page loading, to transform a filter textual
value (read from URL) into a normalized filter option. This is bad,
because triggers unnecessary API calls on page load.

The change consists in splitting concerns between autocompletion and
text mapping to normalized filter option.